+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.41]
+### Changed
+- `Location` accuracy split to horizontal and vertical
+
+### Fixed
+- build problems in the Android module
+
 ## [0.9.40]
 ### Changed
 - `Location` object has completely new structure for meta-data. Changes in clients code needed.